refactor: `validate-at-property.ts` -> `helpers/css-custom-property.ts`
